Reorganize discussion files
