+++ src/features/DiscussionBoard/_components/DiscussionPost/index.scss
